requirements/development.txt
asgiref==3.7.2
astroid==3.0.2
colorama==0.4.6
dill==0.3.7
Django==4.2.9
django-browser-reload==1.12.1
django-neomodel==0.1.1
isort==5.13.2
mccabe==0.7.0
neo4j==5.10.0
neo4j-driver==5.11.0
neobolt==1.7.17
neomodel==5.1.0
platformdirs==4.1.0
pylint==3.0.3
python-decouple==3.8
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2023.4
Unipath==1.1